badash-api/requirements/base.txt
hug
mongoengine
python-slugify
pyjwt
